Add hut as resource, track huts and wood, set cap
Huts are now a resource that can be built for a wood cost. That cost
increases the more huts there are. Huts have a cap of 10 maximum. We may
need to tweak the code in the future to handle fire burning down huts and
appropriately setting the cost. Some renaming is about to be done.