Fix adding huts, append message to textarea, costs
The costs should now be easy to set based on the number of huts.
Incrementing of huts and decrementing of wood works properly. Variable
names should now be appropriate. The user now gets feedback in the
textarea depending on if they have enough wood, can add a hut, or have
reached the limit.